requirements.txt
# Eskişehir Kayaç Parkı - Yapay Zekâ Rehberi
# Dependencies

streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
scikit-learn>=1.3.0
PyPDF2>=3.0.0

# Google Gemini AI
google-genai>=1.0.0
